refactor: extract worker loop from app/__init__.py to core/worker.py
